requirements.txt
atomicwrites==1.4.1
attrs==22.1.0
black==22.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
coverage==6.4.4
cssbeautifier==1.14.5
Django==2.2.16
djangorestframework==3.12.4
djlint==1.9.3
EditorConfig==0.12.3
gunicorn==20.0.4
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
jsbeautifier==1.14.5
Markdown==3.4.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
psycopg2-binary==2.8.6
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
pytest-cov==3.0.0
python-dotenv==0.20.0
pytz==2022.2.1
PyYAML==6.0
regex==2022.8.17
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tqdm==4.64.0
typed-ast==1.5.4
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.11
zipp==3.8.1
